feat: finish project and add readme

+++ test-sps-server/src/User/Services/DeleteUserService.ts
@@ -1,13 +1,19 @@
 import { IUserRepository } from "../Repositories/IUserRepository";
 import { UserDto } from "../Dto/UserDto";
+import { BusinessRuleError } from "../../Errors/BusinessRuleError";
+import { NotFoundError } from "../../Errors/NotFoundError";
 
 export class DeleteUserService {
   constructor(private userRepository: IUserRepository) {}
 
-  public async execute(data: UserDto) {
+  public async execute(data: UserDto, currentUserId: string) {
+    if (data.id === currentUserId) {
+      throw new BusinessRuleError("Usuário logado não pode se deletar!");
+    }
+
     const exists = await this.userRepository.findById(data.id);
     if (!exists) {
-      throw new Error("Usuário não existe!");
+      throw new NotFoundError("Usuário não existe!");
     }
 
     return this.userRepository.delete(data.id);
